requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
PyMuPDF==1.24.10
openpyxl==3.1.5
pandas==2.2.3
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.3.0
googletrans==4.0.2
azure-ai-formrecognizer==3.3.3
azure-keyvault-secrets==4.9.0
azure-identity
cachetools==5.5.1
azure-storage-blob==12.24.1
openai==1.63.1
asyncio==3.4.3
aiohttp==3.11.16
xlrd==1.2.0
mistralai==1.7.0
pyarrow
pinecone==8.0.0
selenium==4.27.1